chore(core): move soul components to root folder /vibes/soul
