+++ core/tsconfig.json
@@ -38,7 +38,8 @@
     ],
     "baseUrl": ".",
     "paths": {
-      "~/*": ["./*"]
+      "@/vibes/*": ["./vibes/*"],
+      "~/*": ["./*"],
     },
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
   },
