+++ core/vibes/soul/components/badge/index.tsx
