+++ core/vibes/soul/components/carousel/index.tsx
