+++ core/vibes/soul/components/footer/index.tsx
