+++ core/vibes/soul/components/footer/payment-icons.tsx
