+++ core/vibes/soul/components/footer/social-icons.tsx
