+++ core/vibes/soul/components/inline-email-form/index.tsx
