+++ core/vibes/soul/components/product-card/price.tsx
